+++ clang/lib/Basic/LangOptions.cpp
@@ -240,3 +240,54 @@ LLVM_DUMP_METHOD void FPOptionsOverride::dump() {
 #include "clang/Basic/FPOptions.def"
   llvm::errs() << "\n";
 }
+
+AtomicOptionsOverride
+AtomicOptions::getChangesSlow(const AtomicOptions &Base) const {
+  AtomicOptions::storage_type OverrideMask = 0;
+#define OPTION(NAME, STR, TYPE, WIDTH, PREVIOUS)                               \
+  if (get##NAME() != Base.get##NAME())                                         \
+    OverrideMask |= NAME##Mask;
+#include "clang/Basic/AtomicOptions.def"
+  return AtomicOptionsOverride(*this, OverrideMask);
+}
+
+LLVM_DUMP_METHOD void AtomicOptions::dump() {
+#define OPTION(NAME, STR, TYPE, WIDTH, PREVIOUS)                               \
+  llvm::errs() << "\n " #NAME " " << get##NAME();
+#include "clang/Basic/AtomicOptions.def"
+  llvm::errs() << "\n";
+}
+
+LLVM_DUMP_METHOD void AtomicOptionsOverride::dump() {
+#define OPTION(NAME, STR, TYPE, WIDTH, PREVIOUS)                               \
+  if (has##NAME##Override())                                                   \
+    llvm::errs() << "\n " #NAME " Override is " << get##NAME##Override();
+#include "clang/Basic/AtomicOptions.def"
+  llvm::errs() << "\n";
+}
+
+static bool isTruthy(StringRef Value) {
+  return Value.equals_insensitive("on") || Value.equals_insensitive("true") ||
+         Value.equals_insensitive("1") || Value.equals_insensitive("yes");
+}
+
+std::optional<AtomicOverrideKind>
+AtomicOptionsOverride::parseAtomicOverrideKey(StringRef Key) {
+  return llvm::StringSwitch<std::optional<AtomicOverrideKind>>(Key)
+#define OPTION(NAME, STR, ...) .Case(STR, AtomicOverrideKind::NAME)
+#include "clang/Basic/AtomicOptions.def"
+      .Default(std::nullopt);
+}
+
+AtomicOptionsOverride::AtomicOptionsOverride(const LangOptions &LO) {
+  for (const auto &Setting : LO.AtomicOptionsAsWritten) {
+    SmallVector<StringRef, 2> KeyValue;
+    StringRef(Setting).split(KeyValue, ":");
+    // Assuming option string has been checked elsewhere and is valid.
+    assert(KeyValue.size() == 2 && "Invalid atomic option format");
+    if (auto Kind = parseAtomicOverrideKey(KeyValue[0])) {
+      bool IsEnabled = isTruthy(KeyValue[1]);
+      setAtomicOverride(*Kind, IsEnabled);
+    }
+  }
+}